[refactor] vuex 구조 변경
