+++ src/sgs-client/src/components/LoginForm.vue
@@ -80,6 +80,7 @@
 </template>
 
 <script>
+import { mapActions } from "vuex";
 import { validateEmail } from "../utils/validation.js";
 export default {
   data() {
@@ -102,13 +103,14 @@ export default {
     },
   },
   methods: {
+    ...mapActions("auth", ["LOGIN"]),
     async submitForm() {
       try {
         const userData = {
           email: this.id,
           password: this.pwd,
         };
-        await this.$store.dispatch("LOGIN", userData);
+        await this.LOGIN(userData);
         this.$router.push("/channels/@me");
       } catch (err) {
         this.logMessage = "로그인에 실패하셨습니다.";
