+++ src/sgs-client/src/store/index.js
@@ -1,37 +1,10 @@
 import Vue from "vue";
 import Vuex from "vuex";
-import mutations from "./mutations.js";
-import actions from "./actions.js";
-import {
-  getAccessAuthToCookie,
-  getRefreshAuthToCookie,
-  getUserEmailToCookie,
-  getUserNickNameToCookie,
-  getUserCodeToCookie,
-} from "../utils/cookies";
+/* import auth from "./module/auth.js";
+import friends from "./module/friends.js";
+import server from "./module/server.js"; */
+import modules from "./module";
 Vue.use(Vuex);
 
-export default new Vuex.Store({
-  state: {
-    email: getUserEmailToCookie() || "",
-    nickname: getUserNickNameToCookie() || "",
-    code: getUserCodeToCookie() || "",
-    accesstoken: getAccessAuthToCookie() || "",
-    refreshtoken: getRefreshAuthToCookie() || "",
-    createchannel: false,
-    friendsstatemenu: "online",
-    friendsonline: [],
-    friendsall: [],
-    friendsreceivedwaiting: [],
-    friendsreceivedwaitingnumber: 0,
-    friendssendwaiting: [],
-    friendsblocked: [],
-  },
-  getters: {
-    isLogin(state) {
-      return state.email !== "";
-    },
-  },
-  mutations,
-  actions,
-});
+const store = new Vuex.Store({ modules });
+export default store;
