+++ src/sgs-client/src/store/module/index.js
@@ -1,0 +1,5 @@
+import auth from "./auth.js";
+import friends from "./friends.js";
+import server from "./server.js";
+
+export default { auth, friends, server };
