+++ src/sgs-client/src/store/module/server.js
@@ -1,0 +1,13 @@
+const server = {
+  namespaced: true,
+  state: {
+    createchannel: false,
+  },
+  mutations: {
+    setCreateChannel(state, createchannel) {
+      state.createchannel = createchannel;
+    },
+  },
+};
+
+export default server;
